+++ packages/api-server/.gitignore
@@ -1,1 +1,3 @@
+/.venv/
+/api_server.egg-info
 /static/cache/
