+++ packages/api-server/README.md
@@ -3,13 +3,13 @@
 Let's create a virtual environment named `venv` for all this Python magic:
 ```
 sudo apt install python3-venv
-python3 -m venv venv
+python3 -m venv --system-site-packages .venv
 ```
 
-Now let's activate the Python virtual environment so we can install all sorts of things into it without breaking our system:
+Now let's activate the Python virtual environment and install all sorts of things into it without breaking our system:
 ```
-. venv/bin/activate
-pip install flask python-dotenv
+. .venv/bin/activate
+pip3 install -e .
 ```
 
 # Run development server
