--- packages/api-server/app.py
@@ -1,97 +1,0 @@
-import base64
-import hashlib
-import os
-import signal
-import threading
-from urllib.parse import urljoin
-
-import rclpy
-from building_map_msgs.msg import AffineImage, BuildingMap, Level
-from flask import Flask
-from rclpy.node import Node
-from rosidl_runtime_py.convert import message_to_ordereddict
-
-
-class MainNode(Node):
-    def __init__(self):
-        super().__init__('rmf_api_server')
-        self.building_map = None
-
-        def _on_building_map(building_map: BuildingMap):
-            '''
-            1. Converts a `BuildingMap` message to an ordered dict.
-            2. Saves the images into `static/cache/{map_name}`.
-            3. Change the `AffineImage` `data` field to the url of the image.
-            '''
-            app.logger.info(f'received new building map "{building_map.name}"')
-            # this is inefficient as it converts a large bytearray to a list of numbers, but
-            # there isn't any better alternatives
-            self.building_map = message_to_ordereddict(building_map)
-            os.makedirs(
-                f'static/cache/{building_map.name}', exist_ok=True)
-            for i in range(len(building_map.levels)):
-                level: Level = building_map.levels[i]
-                for j in range(len(level.images)):
-                    image: AffineImage = level.images[j]
-                    # look at non-crypto hashes if we need more performance
-                    sha1_hash = hashlib.sha1()
-                    sha1_hash.update(image.data)
-                    fingerprint = base64.b32encode(
-                        sha1_hash.digest()).lower().decode()
-                    filepath = f'static/cache/{building_map.name}/{level.name}-{image.name}-{fingerprint}.{image.encoding}'
-                    # serve this with a more efficient webserver like nginx or a cdn if we need
-                    # more performance
-                    with open(filepath, 'bw') as f:
-                        f.write(image.data)
-                        app.logger.info(f'saved map image to "{filepath}"')
-                    self.building_map['levels'][i]['images'][j]['data'] = \
-                        urljoin(app.config['APPLICATION_ROOT'], filepath)
-            app.logger.info(f'updated building map "{building_map.name}"')
-
-        self. building_map_sub = self.create_subscription(
-            BuildingMap,
-            'map',
-            _on_building_map,
-            rclpy.qos.QoSProfile(
-                history=rclpy.qos.HistoryPolicy.KEEP_LAST,
-                depth=1,
-                reliability=rclpy.qos.ReliabilityPolicy.RELIABLE,
-                durability=rclpy.qos.DurabilityPolicy.TRANSIENT_LOCAL,
-            )
-        )
-
-
-def ros2_thread(node):
-    print('entering ros2 thread')
-    rclpy.spin(node)
-    print('leaving ros2 thread')
-
-
-def sigint_handler(signal, frame):
-    '''
-    SIGINT handler
-
-    We have to know when to tell rclpy to shut down, because
-    it's in a child thread which would stall the main thread
-    shutdown sequence. So we use this handler to call
-    rclpy.shutdown() and then call the previously-installed
-    SIGINT handler for Flask
-    '''
-    rclpy.shutdown()
-    if prev_sigint_handler is not None:
-        prev_sigint_handler(signal)
-
-
-rclpy.init(args=None)
-ros2_node = MainNode()
-app = Flask(__name__)
-if 'RMF_API_APP_ROOT' in os.environ:
-    app.config['APPLICATION_ROOT'] = os.environ['RMF_API_ROOT']
-
-threading.Thread(target=ros2_thread, args=[ros2_node]).start()
-prev_sigint_handler = signal.signal(signal.SIGINT, sigint_handler)
-
-
-@app.route('/building_map')
-def get_building_map():
-    return ros2_node.building_map
