+++ packages/api-server/run.sh
@@ -1,2 +1,6 @@
 #!/bin/bash
+
+export FLASK_APP=api_server.app
+export FLASK_ENV=development
+
 flask run --eager-loading --port 5001
